:lipstick: Reduce page jump
When pre-rendering, all translated text is left blank and then filled in by JS. This causes the Main Menu's heading to be small when there is no text, then make everything jump down when it loads. Make it at least as big as one line of text to solve this.