:bug: Fix omitted fields bug
Since the first player ID is always 0, any field containing this ID is determined to be omitted since it is the zero-value of that field in go. Since explicit is better than implicit, we should handle this on the client. It is simple to make the first PID be 1 and increment from there.