:bug: Use AfterFunc over Sleep in backend
If the melody session is closed before messages are sent, clients won't receive them. A delay allows the messages to be sent. AfterFunc is used instead of Sleep to prevent locking issues. Because AfterFunc allows the function to return, this fixes a locking issue in Game.handleGameOver when g.End tries to aquire the lock while it is possessed by g.handleGameOver()'s caller, leading to an infinite hang. Another solution would be to use Sleep in conjunction with go g.End() but AfterFunc seems more idiomatic than spawning another goroutine.